c/controller_backend: move internal structs to .cc
